playerViewからrender部分外した
